moved dataclasses to a new file
